// src/api/models/counselor-page-res.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CounselorListItem } from './counselor-list-item';
// May contain unused imports in some cases
// @ts-ignore
import type { PageRes } from './page-res';

/**
 * 
 * @export
 * @interface CounselorPageRes
 */
export interface CounselorPageRes {
    /**
     * 
     * @type {Array<CounselorListItem>}
     * @memberof CounselorPageRes
     */
    'counselors'?: Array<CounselorListItem>;
    /**
     * 
     * @type {PageRes}
     * @memberof CounselorPageRes
     */
    'pageInfo'?: PageRes;
}